Move some key identification code that will be useful in the cgame to src/common.
