+++ src/engine/client/key_identification.cpp
@@ -32,8 +32,6 @@ Maryland 20850 USA.
 ===========================================================================
 */
 
-// Object and string representations of keyboard keys.
-
 #include "common/Common.h"
 
 #include "key_identification.h"
@@ -43,15 +41,12 @@ Maryland 20850 USA.
 
 namespace Keyboard {
 
-const Key Key::NONE = Key();
-const Key Key::CONSOLE = Key(Key::Kind::CONSOLE, 0);
-
-Key Key::FromScancode(int scancode) {
+Key KeyFromScancode(int scancode) {
 	if (scancode > static_cast<int>(SDL_SCANCODE_UNKNOWN)
 		&& scancode < static_cast<int>(SDL_NUM_SCANCODES)) {
-		return Key(Kind::SCANCODE, scancode);
+		return Key(Key::Kind::SCANCODE, scancode);
 	}
-	return NONE;
+	return Key::NONE;
 }
 
 // Returns the scancode of a char for a QWERTY keyboard
@@ -114,190 +109,8 @@ char ScancodeToAscii(int sc)
 }
 
 
-struct keyname_t
-{
-	const char *name;
-	Keyboard::Key keynum;
-};
-
-// names not in this list can either be lowercase ascii, or '0xnn' hex sequences
-static const std::unordered_map<Str::StringRef, keyNum_t, Str::IHash, Str::IEqual> keynames
-{
-	{ "TAB",                    keyNum_t::K_TAB          },
-	{ "ENTER",                  keyNum_t::K_ENTER        },
-	{ "ESCAPE",                 keyNum_t::K_ESCAPE       },
-	{ "BACKSPACE",              keyNum_t::K_BACKSPACE    },
-	{ "UPARROW",                keyNum_t::K_UPARROW      },
-	{ "DOWNARROW",              keyNum_t::K_DOWNARROW    },
-	{ "LEFTARROW",              keyNum_t::K_LEFTARROW    },
-	{ "RIGHTARROW",             keyNum_t::K_RIGHTARROW   },
-
-	{ "ALT",                    keyNum_t::K_ALT          },
-	{ "CTRL",                   keyNum_t::K_CTRL         },
-	{ "SHIFT",                  keyNum_t::K_SHIFT        },
-
-	{ "COMMAND",                keyNum_t::K_COMMAND      },
-
-	{ "CAPSLOCK",               keyNum_t::K_CAPSLOCK     },
-
-	{ "F1",                     keyNum_t::K_F1           },
-	{ "F2",                     keyNum_t::K_F2           },
-	{ "F3",                     keyNum_t::K_F3           },
-	{ "F4",                     keyNum_t::K_F4           },
-	{ "F5",                     keyNum_t::K_F5           },
-	{ "F6",                     keyNum_t::K_F6           },
-	{ "F7",                     keyNum_t::K_F7           },
-	{ "F8",                     keyNum_t::K_F8           },
-	{ "F9",                     keyNum_t::K_F9           },
-	{ "F10",                    keyNum_t::K_F10          },
-	{ "F11",                    keyNum_t::K_F11          },
-	{ "F12",                    keyNum_t::K_F12          },
-	{ "F13",                    keyNum_t::K_F13          },
-	{ "F14",                    keyNum_t::K_F14          },
-	{ "F15",                    keyNum_t::K_F15          },
-
-	{ "INS",                    keyNum_t::K_INS          },
-	{ "DEL",                    keyNum_t::K_DEL          },
-	{ "PGDN",                   keyNum_t::K_PGDN         },
-	{ "PGUP",                   keyNum_t::K_PGUP         },
-	{ "HOME",                   keyNum_t::K_HOME         },
-	{ "END",                    keyNum_t::K_END          },
-
-	{ "MOUSE1",                 keyNum_t::K_MOUSE1       },
-	{ "MOUSE2",                 keyNum_t::K_MOUSE2       },
-	{ "MOUSE3",                 keyNum_t::K_MOUSE3       },
-	{ "MOUSE4",                 keyNum_t::K_MOUSE4       },
-	{ "MOUSE5",                 keyNum_t::K_MOUSE5       },
-
-	{ "MWHEELUP",               keyNum_t::K_MWHEELUP     },
-	{ "MWHEELDOWN",             keyNum_t::K_MWHEELDOWN   },
-
-	{ "JOY1",                   keyNum_t::K_JOY1         },
-	{ "JOY2",                   keyNum_t::K_JOY2         },
-	{ "JOY3",                   keyNum_t::K_JOY3         },
-	{ "JOY4",                   keyNum_t::K_JOY4         },
-	{ "JOY5",                   keyNum_t::K_JOY5         },
-	{ "JOY6",                   keyNum_t::K_JOY6         },
-	{ "JOY7",                   keyNum_t::K_JOY7         },
-	{ "JOY8",                   keyNum_t::K_JOY8         },
-	{ "JOY9",                   keyNum_t::K_JOY9         },
-	{ "JOY10",                  keyNum_t::K_JOY10        },
-	{ "JOY11",                  keyNum_t::K_JOY11        },
-	{ "JOY12",                  keyNum_t::K_JOY12        },
-	{ "JOY13",                  keyNum_t::K_JOY13        },
-	{ "JOY14",                  keyNum_t::K_JOY14        },
-	{ "JOY15",                  keyNum_t::K_JOY15        },
-	{ "JOY16",                  keyNum_t::K_JOY16        },
-	{ "JOY17",                  keyNum_t::K_JOY17        },
-	{ "JOY18",                  keyNum_t::K_JOY18        },
-	{ "JOY19",                  keyNum_t::K_JOY19        },
-	{ "JOY20",                  keyNum_t::K_JOY20        },
-	{ "JOY21",                  keyNum_t::K_JOY21        },
-	{ "JOY22",                  keyNum_t::K_JOY22        },
-	{ "JOY23",                  keyNum_t::K_JOY23        },
-	{ "JOY24",                  keyNum_t::K_JOY24        },
-	{ "JOY25",                  keyNum_t::K_JOY25        },
-	{ "JOY26",                  keyNum_t::K_JOY26        },
-	{ "JOY27",                  keyNum_t::K_JOY27        },
-	{ "JOY28",                  keyNum_t::K_JOY28        },
-	{ "JOY29",                  keyNum_t::K_JOY29        },
-	{ "JOY30",                  keyNum_t::K_JOY30        },
-	{ "JOY31",                  keyNum_t::K_JOY31        },
-	{ "JOY32",                  keyNum_t::K_JOY32        },
-
-	{ "AUX1",                   keyNum_t::K_AUX1         },
-	{ "AUX2",                   keyNum_t::K_AUX2         },
-	{ "AUX3",                   keyNum_t::K_AUX3         },
-	{ "AUX4",                   keyNum_t::K_AUX4         },
-	{ "AUX5",                   keyNum_t::K_AUX5         },
-	{ "AUX6",                   keyNum_t::K_AUX6         },
-	{ "AUX7",                   keyNum_t::K_AUX7         },
-	{ "AUX8",                   keyNum_t::K_AUX8         },
-	{ "AUX9",                   keyNum_t::K_AUX9         },
-	{ "AUX10",                  keyNum_t::K_AUX10        },
-	{ "AUX11",                  keyNum_t::K_AUX11        },
-	{ "AUX12",                  keyNum_t::K_AUX12        },
-	{ "AUX13",                  keyNum_t::K_AUX13        },
-	{ "AUX14",                  keyNum_t::K_AUX14        },
-	{ "AUX15",                  keyNum_t::K_AUX15        },
-	{ "AUX16",                  keyNum_t::K_AUX16        },
-
-	{ "KP_HOME",                keyNum_t::K_KP_HOME      },
-	{ "KP_7",                   keyNum_t::K_KP_HOME      },
-	{ "KP_UPARROW",             keyNum_t::K_KP_UPARROW   },
-	{ "KP_8",                   keyNum_t::K_KP_UPARROW   },
-	{ "KP_PGUP",                keyNum_t::K_KP_PGUP      },
-	{ "KP_9",                   keyNum_t::K_KP_PGUP      },
-	{ "KP_LEFTARROW",           keyNum_t::K_KP_LEFTARROW },
-	{ "KP_4",                   keyNum_t::K_KP_LEFTARROW },
-	{ "KP_5",                   keyNum_t::K_KP_5         },
-	{ "KP_RIGHTARROW",          keyNum_t::K_KP_RIGHTARROW},
-	{ "KP_6",                   keyNum_t::K_KP_RIGHTARROW},
-	{ "KP_END",                 keyNum_t::K_KP_END       },
-	{ "KP_1",                   keyNum_t::K_KP_END       },
-	{ "KP_DOWNARROW",           keyNum_t::K_KP_DOWNARROW },
-	{ "KP_2",                   keyNum_t::K_KP_DOWNARROW },
-	{ "KP_PGDN",                keyNum_t::K_KP_PGDN      },
-	{ "KP_3",                   keyNum_t::K_KP_PGDN      },
-	{ "KP_ENTER",               keyNum_t::K_KP_ENTER     },
-	{ "KP_INS",                 keyNum_t::K_KP_INS       },
-	{ "KP_0",                   keyNum_t::K_KP_INS       },
-	{ "KP_DEL",                 keyNum_t::K_KP_DEL       },
-	{ "KP_PERIOD",              keyNum_t::K_KP_DEL       },
-	{ "KP_SLASH",               keyNum_t::K_KP_SLASH     },
-	{ "KP_MINUS",               keyNum_t::K_KP_MINUS     },
-	{ "KP_PLUS",                keyNum_t::K_KP_PLUS      },
-	{ "KP_NUMLOCK",             keyNum_t::K_KP_NUMLOCK   },
-	{ "KP_STAR",                keyNum_t::K_KP_STAR      },
-	{ "KP_EQUALS",              keyNum_t::K_KP_EQUALS    },
-
-	{ "PAUSE",                  keyNum_t::K_PAUSE        },
-
-	{ "WINDOWS",                keyNum_t::K_SUPER        },
-	{ "COMPOSE",                keyNum_t::K_COMPOSE      },
-	{ "MODE",                   keyNum_t::K_MODE         },
-	{ "HELP",                   keyNum_t::K_HELP         },
-	{ "PRINT",                  keyNum_t::K_PRINT        },
-	{ "SYSREQ",                 keyNum_t::K_SYSREQ       },
-	{ "SCROLLLOCK",             keyNum_t::K_SCROLLLOCK   },
-	{ "BREAK",                  keyNum_t::K_BREAK        },
-	{ "MENU",                   keyNum_t::K_MENU         },
-	{ "POWER",                  keyNum_t::K_POWER        },
-	{ "EURO",                   keyNum_t::K_EURO         },
-	{ "UNDO",                   keyNum_t::K_UNDO         },
-
-	{ "XBOX360_A",              keyNum_t::K_XBOX360_A              },
-	{ "XBOX360_B",              keyNum_t::K_XBOX360_B              },
-	{ "XBOX360_X",              keyNum_t::K_XBOX360_X              },
-	{ "XBOX360_Y",              keyNum_t::K_XBOX360_Y              },
-	{ "XBOX360_LB",             keyNum_t::K_XBOX360_LB             },
-	{ "XBOX360_RB",             keyNum_t::K_XBOX360_RB             },
-	{ "XBOX360_START",          keyNum_t::K_XBOX360_START          },
-	{ "XBOX360_GUIDE",          keyNum_t::K_XBOX360_GUIDE          },
-	{ "XBOX360_LS",             keyNum_t::K_XBOX360_LS             },
-	{ "XBOX360_RS",             keyNum_t::K_XBOX360_RS             },
-	{ "XBOX360_BACK",           keyNum_t::K_XBOX360_BACK           },
-	{ "XBOX360_LT",             keyNum_t::K_XBOX360_LT             },
-	{ "XBOX360_RT",             keyNum_t::K_XBOX360_RT             },
-	{ "XBOX360_DPAD_UP",        keyNum_t::K_XBOX360_DPAD_UP        },
-	{ "XBOX360_DPAD_RIGHT",     keyNum_t::K_XBOX360_DPAD_RIGHT     },
-	{ "XBOX360_DPAD_DOWN",      keyNum_t::K_XBOX360_DPAD_DOWN      },
-	{ "XBOX360_DPAD_LEFT",      keyNum_t::K_XBOX360_DPAD_LEFT      },
-	{ "XBOX360_DPAD_RIGHTUP",   keyNum_t::K_XBOX360_DPAD_RIGHTUP   },
-	{ "XBOX360_DPAD_RIGHTDOWN", keyNum_t::K_XBOX360_DPAD_RIGHTDOWN },
-	{ "XBOX360_DPAD_LEFTUP",    keyNum_t::K_XBOX360_DPAD_LEFTUP    },
-	{ "XBOX360_DPAD_LEFTDOWN",  keyNum_t::K_XBOX360_DPAD_LEFTDOWN  },
-};
-
 static const Str::StringRef CHARACTER_BIND_PREFIX = "char:";
 static const Str::StringRef SCANCODE_ASCII_BIND_PREFIX = "hw:";
-
-static const std::unordered_map<char, Str::StringRef> SPECIAL_CHARACTER_NAMES {
-    {';', "SEMICOLON"},
-    {'\\', "BACKSLASH"},
-    {' ', "SPACE"},
-    {'"', "DOUBLEQUOTE"},
-};
 
 static int ParseCharacter(Str::StringRef s)
 {
@@ -319,9 +132,9 @@ static Key KeyFromUnprefixedCharacter(int ch)
 {
     SDL_Scancode sc = SDL_GetScancodeFromKey(static_cast<SDL_Keycode>(ch));
     if (sc != SDL_SCANCODE_UNKNOWN) {
-        return Key::FromScancode(sc);
+        return KeyFromScancode(sc);
     }
-    return Key::FromScancode(AsciiToScancode(ch));
+    return KeyFromScancode(AsciiToScancode(ch));
 }
 
 /*
@@ -344,14 +157,14 @@ Key StringToKey(Str::StringRef str)
     if ( Str::IsIPrefix("0x", str) )
     {
         int n = Com_HexStrToInt( str.c_str() );
-        return Key::FromScancode( n );
+        return KeyFromScancode( n );
     }
 
     // Physical key by QWERTY location, from ascii char
     size_t prefixLen = SCANCODE_ASCII_BIND_PREFIX.size();
     if ( Str::IsIPrefix( SCANCODE_ASCII_BIND_PREFIX, str ) )
     {
-        return Key::FromScancode( AsciiToScancode( ParseCharacter ( str.substr( prefixLen ) ) ) );
+        return KeyFromScancode( AsciiToScancode( ParseCharacter ( str.substr( prefixLen ) ) ) );
     }
 
     // char:X forces a virtual key-based rather than scancode binding to be used for some (Unicode) char
@@ -367,26 +180,6 @@ Key StringToKey(Str::StringRef str)
     }
 
     return Key::NONE;
-}
-
-std::string CharToString(int ch)
-{
-    auto it = SPECIAL_CHARACTER_NAMES.find(ch);
-    if (it != SPECIAL_CHARACTER_NAMES.end()) {
-        return it->second;
-    }
-    return Q_UTF8_Encode(ch);
-}
-
-std::string KeynumToString(keyNum_t keynum)
-{
-    for (auto& kv : keynames) {
-        if (kv.second == keynum) {
-            return kv.first;
-        }
-    }
-    Log::Warn("Unknown keynum %d", Util::ordinal(keynum));
-    return "<UNKNOWN KEYNUM>";
 }
 
 std::string KeyToString(Key key)
